+++ changelog.md
@@ -1,11 +1,13 @@
 ## Changelog
 
-### v1.5.1 -
+### v1.5.1 - Jan 18, 2019
+- `Add` Feature: Get anime/manga recommendation
 - `Add` Feature: Additional anime/manga recommendation
 - `Add` Feature: Get all anime/manga recommendation list
 - `Fix` Missing recommendation in `getInfo()`
 
 ### v1.5.0 - Jan 15, 2019
+- `Add` Feature: Get anime/manga review
 - `Add` Feature: Additional anime/manga review
 - `Add` Feature: Get all anime/manga/best-voted review list
 - `Fix` Missing review in `getInfo()`